# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZK-VOTE Relayer API",
    "version": "1.0.0",
    "description": "Backend relayer for anonymous voting on Stellar/Soroban. Generated from route definitions and Zod validation schemas — see backend/API.md for prose docs and GET /api-docs for interactive documentation."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "RelayerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relayer-Auth",
        "description": "Shared relayer auth token (also accepted as `Authorization: Bearer <token>`). There is currently one shared token per deployment, not per-user credentials."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Unauthorized"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/admin/audit-log": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Paginated, hash-chain-verifiable audit log review (admin only)",
        "description": "Rate limit: queryLimiter.",
        "security": [
          {
            "RelayerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "50"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "0"
            },
            "required": false,
            "name": "offset",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "vote_relay"
            },
            "required": false,
            "name": "action",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "cef"
              ]
            },
            "required": false,
            "name": "format",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "verify",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "logs": [],
                  "total": 0,
                  "limit": 50,
                  "offset": 0
                }
              }
            }
          },
          "401": {
            "description": "Error (HTTP 401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error (HTTP 500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/sbt-transfer-attempts": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Review flagged membership-SBT transfer/approval attempts for a DAO (admin only)",
        "description": "Rate limit: queryLimiter.",
        "security": [
          {
            "RelayerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "1"
            },
            "required": true,
            "name": "daoId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "50"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "0"
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "daoId": 1,
                  "attempts": [],
                  "total": 0,
                  "limit": 50,
                  "offset": 0
                }
              }
            }
          },
          "400": {
            "description": "Error (HTTP 400)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error (HTTP 401)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error (HTTP 500)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}